Minimal changes to get all tests to pass - in the Java implementation.
I've captured the thought-process in new file NOTES.md (which would normally be
put as comments/subtasks in the issue-tracking system, if this was "for real").

I haven't changed any fundamental representations of things, I've just made small tweaks to
the existing structure, and filled out implementations for the methods that had been implemented
as just NotImplemented stubs.

The only API change I've made is to change the values of MAX_BOARD_WIDTH and MAX_BOARD_HEIGHT
to 8, rather than 7. It's possible that they were deliberately set to 7 as a "non-standard board
size", but I doubt it (especially as the images in the links from the README.md file all point to
pictures of standard-size 8x8 boards).

The code still needs a fair amount of cleanup, refactoring, and comments, and needs some more
tests - but at least the existing tests are now all passing.