Added another test to check coordinates match between pieces and board, after a piece is moved.
And in doing so, unearthed (and fixed) a bug in that.

Also extended the ChessBoard public API to include a new pieceAt() method, so you can ask the
board for the piece (currently a Pawn or null) that's at a particular position. Needed for tests
for now, but I'm expecting we'll need this more generally soon anyway.

Also changed the name of some variables inside the tests to make it clearer what they're for.