Refactored Piece (and Pawn) class to make it so its constructor takes no arguments.
It previously required a PieceColor parameter, but that was redundant as the ChessBoard
API requires that a piece be allocated a colour when it is placed on the board, anyway.

Also added a check (and test) to ensure that a piece that's already on a board can't be
added to a board. (You can only add a brand-new pristine piece now).