Added new class ChessGame to represent a whole game (a board and all of its pieces).
This takes on some of the management responsibilities of how to get new pieces onto the board,
and abstracts a bit in a way that will help testing, and will make it easier to add support for
additional piece types.

Also added a new method to the ChessBoard class, to support this, for getting hold of a
collection of all of the pieces currently placed on a board.

Did some other minor refactoring too.